refactor(lib): rename lib/state.svelte to lib/global-state.svelte to avoid naming conflicts
