+++ src/lib/svelte/global-state.svelte.ts
